1. 文件解密流程更新

+++ src/entity/Message.go
@@ -18,10 +18,10 @@ type Message struct {
 }
 
 type MessageData struct {
-	Type     int     `json:"type"`     //1-文本 2-图片 3-语音 4-视频 5-文件
-	Content  string  `json:"content"`  //消息内容
-	Width    int     `json:"width"`    //图片宽
-	Height   int     `json:"height"`   //图片高
-	Size     int     `json:"size"`     //文件大小
-	Duration float64 `json:"duration"` //语音时长
+	Type     int    `json:"type"`     //1-文本 2-图片 3-语音 4-视频 5-文件
+	Content  string `json:"content"`  //消息内容
+	Width    int    `json:"width"`    //图片宽
+	Height   int    `json:"height"`   //图片高
+	Size     int    `json:"size"`     //文件大小
+	Duration int    `json:"duration"` //语音时长
 }
